@KaelWD 2018-09-05
Absolute legend, I've been thinking about doing this for a while but didn't know where to start. 

> Faster build speed

Full build speed doesn't matter, we do need fast rebuild when watching though (and working HMR/sourcemaps). Also ideally we wouldn't still need a separate webpack config for `yarn dev`, but that might have to be a combination of webpack-dev-middleware and rollup watch so HMR still works.