@DanSnow 2018-09-05
I got it. But rollup is not supporting HMR and it only fit for building release.
Maybe I can also try to speed up the webpack build.
I think this should be a new PR, right?